Ignore -N option if accurate file layout is specified
This patch ignores any archive name, (specified with the -N
command line option) if the 'accurate' file layout is in
use.

Previously, if the accurate file layout was used in conjuction
with the -N option, rotter would not include the expected
minutes, seconds and hundredths of a second filename parts:

$ ./rotter -L accurate -N archive /tmp
[INFO]   Sun Jul  1 19:20:27 2012  JACK client registered as 'rotter'.
[INFO]   Sun Jul  1 19:20:27 2012  Opening new archive file for ringbuffer A: /tmp/2012-07-01/archive-2012-07-01-19.aiff

$ ./rotter -L accurate  /tmp
[INFO]   Sun Jul  1 19:20:50 2012  JACK client registered as 'rotter'.
[INFO]   Sun Jul  1 19:20:50 2012  Opening new archive file for ringbuffer A: /tmp/2012-07-01/2012-07-01-19-20-50-32.aiff

This patch fixes this inconsistent and undocumented behaviour.